requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web scraping
scrapy==2.11.0
scrapy-playwright==0.0.34
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis & caching
redis==5.0.1
hiredis==2.2.3

# API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2

# CLI
typer[all]==0.9.0
rich==13.7.0

# Task queue
celery==5.3.4
flower==2.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Monitoring & logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Development tools
black==23.12.0
ruff==0.1.8
mypy==1.7.1
pre-commit==3.6.0
ipython==8.18.1

# Data validation
great-expectations==0.18.6
jsonschema==4.20.0

# Utilities
httpx==0.25.2
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.27